Made the Buttons into a single file, Added them to the nav bar. They link to the correct page.
