--- dne-frontend/src/assets/button_bar/Portfolio.tsx
@@ -1,18 +1,0 @@
-import {VNode} from 'preact'
-
-
-const PortfolioIcon = (props: any): VNode => (
-    <svg
-        width="60"
-        height="59"
-        viewBox="0 0 60 59"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-    {...props}
-    >
-        <path d="M55 31C55 44.8071 43.8071 56 30 56C16.1929 56 5 44.8071 5 31C5 17.1929 16.1929 6 30 6C43.8071 6 55 17.1929 55 31Z" fill="white"/>
-        <path d="M53.025 39.0629C51.4346 42.7614 48.947 46.0205 45.7797 48.5552C42.6124 51.09 38.8619 52.8232 34.8561 53.6034C30.8503 54.3835 26.7111 54.1869 22.8004 53.0306C18.8897 51.8744 15.3266 49.7937 12.4225 46.9705C9.51854 44.1473 7.36204 40.6676 6.14159 36.8355C4.92114 33.0034 4.6739 28.9357 5.4215 24.988C6.16909 21.0403 7.88874 17.3328 10.4301 14.1895C12.9715 11.0463 16.2572 8.56311 20 6.95704M55 29.5C55 26.2716 54.3534 23.0749 53.097 20.0923C51.8406 17.1097 49.9991 14.3997 47.6777 12.1169C45.3562 9.83414 42.6002 8.02335 39.5671 6.78792C36.5339 5.55249 33.283 4.91663 30 4.91663V29.5H55Z" stroke="#1E1E1E" stroke-width="4" stroke-linecap="round" stroke-linejoin="round"/>
-    </svg>
-    )
-
-export {PortfolioIcon}
